+++ R/regress.R
@@ -94,246 +94,13 @@ regressModuleUI <- function(id) {
     ),
     checkboxInput(ns("regressUI_subcheck"), "Sub-group analysis"),
     uiOutput(ns("regressUI_subvar")),
-    uiOutput(ns("regressUI_subval"))
+    uiOutput(ns("regressUI_subval")),
+    checkboxInput(ns("step_check"), "Stepwise variable selection"),
+    uiOutput(ns("step_direction")),
+    uiOutput(ns("step_scope"))
   )
 }
 
-
-
-
-
-
-#' @title regressModule: Shiny modulde server for linear regression.
-#' @description Shiny modulde server for linear regression.
-#' @param input input
-#' @param output output
-#' @param session session
-#' @param data data
-#' @param data_label data label
-#' @param data_varStruct List of variable structure, Default: NULL
-#' @param nfactor.limit nlevels limit in factor variable, Default: 10
-#' @param design.survey survey data. default: NULL
-#' @param default.unires Set default independent variables using univariate analysis, Default: T
-#' @param limit.unires Change to default.unires = F if number of independent variables > limit.unires, Default: 20
-#' @return Shiny modulde server for linear regression.
-#' @details Shiny modulde server for linear regression.
-#' @examples
-#' library(shiny);library(DT);library(data.table);library(jstable)
-#' ui <- fluidPage(
-#'    sidebarLayout(
-#'    sidebarPanel(
-#'      regressModuleUI("linear")
-#'    ),
-#'    mainPanel(
-#'      DTOutput("lineartable")
-#'    )
-#'  )
-#')
-#'
-#' server <- function(input, output, session) {
-#'
-#'   data <- mtcars
-#'   data.label <- jstable::mk.lev(mtcars)
-#'
-#'   out_linear <- callModule(regressModule, "linear", data = data, data_label = data.label,
-#'                            data_varStruct = NULL)
-#'
-#'   output$lineartable <- renderDT({
-#'     datatable(out_linear()$table, rownames=T, caption = out_linear()$caption)
-#'   })
-#'}
-#' @rdname regressModule
-#' @export
-#' @import shiny
-#' @importFrom data.table data.table .SD :=
-#' @importFrom labelled var_label<-
-#' @importFrom stats glm as.formula model.frame model.matrix
-#' @importFrom epiDisplay regress.display
-#' @importFrom jstable LabelepiDisplay
-#' @importFrom purrr map_lgl
-
-regressModule <- function(input, output, session, data, data_label, data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL, default.unires = T, limit.unires = 20) {
-
-  ## To remove NOTE.
-  level <- val_label <- variable <- NULL
-
-  if (is.null(data_varStruct)){
-    data_varStruct = list(variable = names(data))
-  }
-
-  if (!("data.table" %in% class(data))) {data = data.table(data)}
-  if (!("data.table" %in% class(data_label))) {data_label = data.table(data_label)}
-
-  factor_vars <- names(data)[data[, lapply(.SD, class) %in% c("factor", "character")]]
-  #data[, (factor_vars) := lapply(.SD, as.factor), .SDcols= factor_vars]
-  factor_list <- mklist(data_varStruct, factor_vars)
-
-  conti_vars <- setdiff(names(data), factor_vars)
-  if (!is.null(design.survey)){
-    conti_vars <- setdiff(conti_vars, c(names(design.survey$allprob), names(design.survey$strata), names(design.survey$cluster)))
-  }
-  conti_list <- mklist(data_varStruct, conti_vars)
-
-  nclass_factor <- unlist(data[, lapply(.SD, function(x){length(levels(x))}), .SDcols = factor_vars])
-
-  group_vars <- factor_vars[nclass_factor >=2 & nclass_factor <= nfactor.limit & nclass_factor < nrow(data)]
-  group_list <- mklist(data_varStruct, group_vars)
-
-  except_vars <- factor_vars[nclass_factor > nfactor.limit | nclass_factor == 1 | nclass_factor == nrow(data)]
-
-
-  output$dep <- renderUI({
-    tagList(
-      selectInput(session$ns("dep_vars"), "Dependent variable",
-                  choices = conti_list, multiple = F,
-                  selected = conti_vars[1]
-      )
-    )
-  })
-
-
-  output$indep <- renderUI({
-    req(!is.null(input$dep_vars))
-    if (is.null(design.survey)){
-      vars <- setdiff(setdiff(names(data), except_vars),  input$dep_vars)
-      if (default.unires){
-        varsIni <- sapply(vars,
-                          function(v){
-                            forms <- as.formula(paste(input$dep_vars, "~", v))
-                            coef <- tryCatch(summary(glm(forms, data = data))$coefficients, error = function(e){return(NULL)})
-                            #coef <- summary(glm(forms, data = data()))$coefficients
-                            sigOK <- ifelse(is.null(coef), F, !all(coef[-1, "Pr(>|t|)"] > 0.05))
-                            return(sigOK)
-                          })
-        if (length(varsIni[varsIni == T]) > limit.unires) {varsIni <- c(T, rep(F, length(vars) -1))}
-      } else{
-        varsIni <- c(T, rep(F, length(vars) -1))
-      }
-
-    } else{
-      vars <- setdiff(setdiff(names(data), except_vars),  c(input$dep_vars, names(design.survey$allprob), names(design.survey$strata), names(design.survey$cluster)))
-      if (default.unires){
-        varsIni <- sapply(vars,
-                          function(v){
-                            forms <- as.formula(paste(input$dep_vars, "~", v))
-                            coef <- tryCatch(summary(survey::svyglm(forms, design = design.survey))$coefficients, error = function(e){return(NULL)})
-                            sigOK <- ifelse(is.null(coef), F, !all(coef[-1, "Pr(>|t|)"] > 0.05))
-                            return(sigOK)
-                          })
-        if (length(varsIni[varsIni == T]) > limit.unires) {varsIni <- c(T, rep(F, length(vars) -1))}
-      } else{
-        varsIni <- c(T, rep(F, length(vars) -1))
-      }
-
-    }
-
-
-    tagList(
-      selectInput(session$ns("indep_vars"), "Independent variables",
-                  choices = mklist(data_varStruct, vars), multiple = T,
-                  selected = vars[varsIni]
-      )
-    )
-  })
-
-  observeEvent(input$regressUI_subcheck, {
-    output$regressUI_subvar <- renderUI({
-      req(input$regressUI_subcheck == T)
-      factor_vars <- names(data)[data[, lapply(.SD, class) %in% c("factor", "character")]]
-      factor_subgroup <- setdiff(factor_vars, c(input$dep_vars, input$indep_vars))
-      factor_subgroup_list <- mklist(data_varStruct, factor_subgroup)
-      validate(
-        need(length(factor_subgroup) > 0 , "No factor variable for sub-group analysis")
-      )
-
-      tagList(
-        selectInput(session$ns("subvar_regress"), "Sub-group variable",
-                    choices = factor_subgroup_list, multiple = F,
-                    selected = factor_subgroup[1])
-      )
-    })
-
-    output$regressUI_subval <- renderUI({
-      req(input$regressUI_subcheck == T)
-      req(input$subvar_regress)
-      selectInput(session$ns("subval_regress"), "Sub-group value",
-                  choices = data_label[variable == input$subvar_regress, level], multiple = F,
-                  selected = data_label[variable == input$subvar_regress, level][1])
-    })
-  })
-
-
-
-
-
-
-  out <- reactive({
-    data.regress <- data
-    if(input$regressUI_subcheck == T){
-      req(input$subvar_regress)
-      data.regress <- data.regress[get(input$subvar_regress) == input$subval_regress, ]
-    }
-    y <- input$dep_vars
-    xs <- input$indep_vars
-    validate(
-      need(!is.null(input$indep_vars) , "Please select at least 1 variable")
-    )
-    form = as.formula(paste(y, "~", paste(xs, collapse = " + "), sep = " "))
-    mf <- model.frame(form, data.regress)
-    validate(
-      need(nrow(mf) > 0, paste("No complete data due to missingness. Please remove some variables from independent variables"))
-    )
-
-    lgl.1level <- purrr::map_lgl(mf, ~length(unique(.x)) == 1)
-    validate(
-      need(sum(lgl.1level) == 0, paste(paste(names(lgl.1level)[lgl.1level], collapse =" ,"), "has(have) a unique value. Please remove that from independent variables"))
-    )
-
-    ## Warning when multicollinearity
-    warn <- NULL
-    mm <- model.matrix(form, data.regress)
-    zz <- tryCatch(solve(t(as.matrix(mm)) %*% as.matrix(mm)), error = function(e) e)
-    if ("error" %in% class(zz)) {
-      warn <- "There are strongly correlated independent values. Please see correlation in `Plot- Scatter plot` tab"
-    }
-
-
-    if (is.null(design.survey)){
-      res.linear <- glm(form, data = data.regress)
-      tb.linear <- jstable::glmshow.display(res.linear, decimal = input$decimal)
-      cap.linear <- paste("Linear regression predicting ", data_label()[variable == y, var_label][1], sep="")
-      if(input$regressUI_subcheck == T){
-        cap.linear <- paste(cap.linear, " - ", data_label()[variable == input$subvar_regress, var_label][1], ": ", data_label()[variable == input$subvar_regress & level == input$subval_regress, val_label], sep = "")
-      }
-      out.linear <- jstable::LabelepiDisplay(tb.linear, label = T, ref = data_label())
-    } else{
-      #validate(
-      #  need(var.weights.survey %in% names(data.regress) , "Weight variable isn't in data-, Please select appropriate weight variable.")
-      #)
-      #data.design <- survey::svydesign(ids = ~ 1, data = data.regress, weights = ~ get(var.weights.survey))
-      data.design <- design.survey
-      if(input$regressUI_subcheck == T){
-        req(input$subvar_regress)
-        data.design <- subset(data.design, get(input$subvar_regress) == input$subval_regress)
-      }
-
-      res.svyglm <- survey::svyglm(form, design = data.design)
-      tb.svyglm <- jstable::svyregress.display(res.svyglm, decimal = input$decimal)
-      cap.linear <- paste("Linear regression predicting ", data_label()[variable == y, var_label][1], "- survey data", sep="")
-      if(input$regressUI_subcheck == T){
-        cap.linear <- paste(cap.linear, " - ", data_label()[variable == input$subvar_regress, var_label][1], ": ", data_label()[variable == input$subvar_regress & level == input$subval_regress, val_label], sep = "")
-      }
-      out.linear <- jstable::LabelepiDisplay(tb.svyglm, label = T, ref = data_label())
-      warn <- NULL
-
-    }
-    return(list(table = out.linear, caption = cap.linear, warning = warn))
-  })
-
-  return(out)
-
-
-}
 
 
 
@@ -381,14 +148,14 @@ regressModule <- function(input, output, session, data, data_label, data_varStru
 #' @import shiny
 #' @importFrom data.table data.table .SD :=
 #' @importFrom labelled var_label<-
-#' @importFrom stats glm as.formula model.frame model.matrix
+#' @importFrom stats glm as.formula model.frame model.matrix step
 #' @importFrom epiDisplay regress.display
 #' @importFrom purrr map_lgl
 
 regressModule2 <- function(input, output, session, data, data_label, data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL, default.unires = T, limit.unires = 20) {
 
   ## To remove NOTE.
-  level <- val_label <- variable <- NULL
+  d2 <- level <- val_label <- variable <- NULL
 
   if (is.null(data_varStruct)){
     data_varStruct <- reactive(list(variable = names(data())))
@@ -536,6 +303,25 @@ regressModule2 <- function(input, output, session, data, data_label, data_varStr
 
   })
 
+  observeEvent(input$step_check, {
+    output$step_direction <- renderUI({
+      req(input$step_check == T)
+      radioButtons(session$ns("step_direction"), "Step direction", choices = c("backward", "forward", "both"), selected = "backward", inline = T)
+    })
+
+    output$step_scope <- renderUI({
+      req(input$step_check == T)
+      req(input$indep_vars)
+      tagList(
+        fluidRow(
+          column(6, selectInput(session$ns("step_lower"), "Lower limit", choices = input$indep_vars, selected = NULL, multiple = T)),
+          column(6, selectInput(session$ns("step_upper"), "Upper limit", choices = input$indep_vars, selected = input$indep_vars, multiple = T))
+        )
+
+      )
+    })
+  })
+
 
 
 
@@ -589,6 +375,17 @@ regressModule2 <- function(input, output, session, data, data_label, data_varStr
       } else{ warn <- NULL}
 
       res.linear <- glm(form, data = data.regress)
+      if (input$step_check == T){
+        validate(
+          need(!is.null(input$step_upper), "Upper limits can't be NULL, please select at least 1 variable."),
+          need(length((setdiff(input$step_lower, input$step_upper))) == 0, "Upper limits must include lower limits. Please add the variables to upper limits")
+        )
+        scope <- lapply(list(input$step_upper, input$step_lower), function(x){
+          as.formula(ifelse(is.null(x), "~1", paste0("~", paste(x, collapse = "+"))))
+        })
+
+        res.linear <- stats::step(glm(form, data = data.regress[complete.cases(data.regress[, .SD, .SDcols = c(y, xs)])]), direction = input$step_direction, scope = list(upper = scope[[1]], lower = scope[[2]]))
+      }
       tb.linear <- jstable::glmshow.display(res.linear, decimal = input$decimal)
       cap.linear <- paste("Linear regression predicting ", label.regress[variable == y, var_label][1], sep="")
       if(input$regressUI_subcheck == T){
@@ -599,6 +396,9 @@ regressModule2 <- function(input, output, session, data, data_label, data_varStr
             cap.linear <- paste(cap.linear, ", ", label.regress[variable == input$subvar_regress[[v]], var_label][1], ": ", paste(input[[paste0("subval_regress", v)]], collapse = "~"), sep = "")
           }
         }
+      }
+      if (input$step_check == T){
+        cap.linear <- paste0(cap.linear, "- stepwise selection")
       }
       out.linear <- jstable::LabelepiDisplay(tb.linear, label = T, ref = label.regress)
     } else{
@@ -635,9 +435,25 @@ regressModule2 <- function(input, output, session, data, data_label, data_varStr
 
 
       res.svyglm <- survey::svyglm(form, design = data.design)
+      if (input$step_check == T){
+        validate(
+          need(!is.null(input$step_upper), "Upper limits can't be NULL, please select at least 1 variable."),
+          need(length((setdiff(input$step_lower, input$step_upper))) == 0, "Upper limits must include lower limits. Please add the variables to upper limits"),
+          need(input$step_direction != "both", "Only forward/bacward directions are supported")
+        )
+        scope <- lapply(list(input$step_upper, input$step_lower), function(x){
+          as.formula(ifelse(is.null(x), "~1", paste0("~", paste(x, collapse = "+"))))
+        })
+
+        d2 <<- na.omit(data.design[, c(y, xs)])
+
+        res.svyglm <- stats::step(survey::svyglm(form, design = d2),
+                                                 direction = input$step_direction, scope = list(upper = scope[[1]], lower = scope[[2]]))
+      }
+
       tb.svyglm <- jstable::svyregress.display(res.svyglm, decimal = input$decimal)
       cap.linear <- paste("Linear regression predicting ", label.regress[variable == y, var_label][1], "- survey data", sep="")
-      if(input$regressUI_subcheck == T){
+      if (input$regressUI_subcheck == T){
         for (v in seq_along(input$subvar_regress)){
           if (input$subvar_regress[[v]] %in% vlist()$factor_vars){
             cap.linear <- paste(cap.linear, ", ", label.regress[variable == input$subvar_regress[[v]], var_label][1], ": ", paste(label.regress[variable == input$subvar_regress[[v]] & level %in% input[[paste0("subval_regress", v)]], val_label], collapse = ", "), sep = "")
@@ -645,6 +461,9 @@ regressModule2 <- function(input, output, session, data, data_label, data_varStr
             cap.linear <- paste(cap.linear, ", ", label.regress[variable == input$subvar_regress[[v]], var_label][1], ": ", paste(input[[paste0("subval_regress", v)]], collapse = "~"), sep = "")
           }
         }
+      }
+      if (input$step_check == T){
+        cap.linear <- paste0(cap.linear, "- stepwise selection")
       }
       out.linear <- jstable::LabelepiDisplay(tb.svyglm, label = T, ref = label.regress)
       warn <- NULL
@@ -659,239 +478,6 @@ regressModule2 <- function(input, output, session, data, data_label, data_varStr
 
 
 }
-
-
-
-
-#' @title logisticModule: Shiny modulde server for logistic regression.
-#' @description Shiny modulde server for logistic regression.
-#' @param input input
-#' @param output output
-#' @param session session
-#' @param data data
-#' @param data_label data label
-#' @param data_varStruct List of variable structure, Default: NULL
-#' @param nfactor.limit nlevels limit in factor variable, Default: 10
-#' @param design.survey survey data. default: NULL
-#' @param default.unires Set default independent variables using univariate analysis, Default: T
-#' @param limit.unires Change to default.unires = F if number of independent variables > limit.unires, Default: 20
-#' @return Shiny modulde server for logistic regression.
-#' @details Shiny modulde server for logistic regression.
-#' @examples
-#' library(shiny);library(DT);library(data.table);library(jstable)
-#' ui <- fluidPage(
-#'    sidebarLayout(
-#'    sidebarPanel(
-#'      regressModuleUI("logistic")
-#'    ),
-#'    mainPanel(
-#'      DTOutput("logistictable")
-#'    )
-#'  )
-#')
-#'
-#' server <- function(input, output, session) {
-#'
-#'   data <- mtcars
-#'   data.label <- jstable::mk.lev(mtcars)
-#'
-#'   out_logistic <- callModule(logisticModule, "logistic", data = data, data_label = data.label,
-#'                            data_varStruct = NULL)
-#'
-#'   output$logistictable <- renderDT({
-#'     datatable(out_logistic()$table, rownames=T, caption = out_logistic()$caption)
-#'   })
-#'}
-#' @rdname logisticModule
-#' @export
-#' @import shiny
-#' @importFrom data.table data.table .SD :=
-#' @importFrom labelled var_label<-
-#' @importFrom stats glm as.formula binomial model.frame quasibinomial
-#' @importFrom epiDisplay logistic.display
-#' @importFrom jstable LabelepiDisplay
-#' @importFrom purrr map_lgl
-
-logisticModule <- function(input, output, session, data, data_label, data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL, default.unires = T, limit.unires = 20) {
-
-  ## To remove NOTE.
-  level <- val_label <- variable <- NULL
-
-  if (is.null(data_varStruct)){
-    data_varStruct = list(variable = names(data))
-  }
-
-  if (!("data.table" %in% class(data))) {data = data.table(data)}
-  if (!("data.table" %in% class(data_label))) {data_label = data.table(data_label)}
-
-  factor_vars <- names(data)[data[, lapply(.SD, class) %in% c("factor", "character")]]
-  #data[, (factor_vars) := lapply(.SD, as.factor), .SDcols= factor_vars]
-  factor_list <- mklist(data_varStruct, factor_vars)
-
-
-  nclass_factor <- unlist(data[, lapply(.SD, function(x){length(levels(x))}), .SDcols = factor_vars])
-
-  factor2_vars <- factor_vars[nclass_factor == 2]
-  if (!is.null(design.survey)){
-    factor2_vars <- setdiff(factor2_vars, c(names(design.survey$strata), names(design.survey$cluster)))
-  }
-  factor2_list <- mklist(data_varStruct, factor2_vars)
-
-  except_vars <- factor_vars[nclass_factor > nfactor.limit | nclass_factor == 1 | nclass_factor == nrow(data)]
-
-
-  output$dep <- renderUI({
-    validate(
-      need(length(factor2_vars) >=1 , "No candidate dependent variable")
-    )
-    tagList(
-      selectInput(session$ns("dep_vars"), "Dependent variable",
-                  choices = factor2_list, multiple = F,
-                  selected = factor2_vars[1]
-      )
-    )
-  })
-
-
-  output$indep <- renderUI({
-    req(!is.null(input$dep_vars))
-
-    if (is.null(design.survey)){
-      vars <- setdiff(setdiff(names(data), except_vars),  input$dep_vars)
-      if (default.unires){
-        varsIni <- sapply(vars,
-                          function(v){
-                            forms <- as.formula(paste(input$dep_vars, "~", v))
-                            coef <- tryCatch(summary(glm(forms, data = data, family = binomial))$coefficients, error = function(e){return(NULL)})
-                            sigOK <- ifelse(is.null(coef), F, !all(coef[-1, 4] > 0.05))
-                            #coef <- summary(glm(forms, data = data(), family = binomial))$coefficients
-                            #sigOK <- !all(coef[-1, 4] > 0.05)
-                            return(sigOK)
-                          })
-        if (length(varsIni[varsIni == T]) > limit.unires) {varsIni <- c(T, rep(F, length(vars) -1))}
-      } else{
-        varsIni <- c(T, rep(F, length(vars)-1))
-      }
-
-    } else{
-      vars <- setdiff(setdiff(names(data), except_vars),  c(input$dep_vars, names(design.survey$allprob), names(design.survey$strata), names(design.survey$cluster)))
-      if (default.unires){
-        varsIni <- sapply(vars,
-                          function(v){
-                            forms <- as.formula(paste(input$dep_vars, "~", v))
-                            coef <- tryCatch(summary(survey::svyglm(forms, design = design.survey, family = binomial))$coefficients, error = function(e){return(NULL)})
-                            sigOK <- ifelse(is.null(coef), F, !all(coef[-1, 4] > 0.05))
-                            return(sigOK)
-                          })
-        if (length(varsIni[varsIni == T]) > limit.unires) {varsIni <- c(T, rep(F, length(vars) -1))}
-      } else{
-        varsIni <- c(T, rep(F, length(vars)-1))
-      }
-
-    }
-
-
-    tagList(
-      selectInput(session$ns("indep_vars"), "Independent variables",
-                  choices = mklist(data_varStruct, vars), multiple = T,
-                  selected = vars[varsIni]
-      )
-    )
-  })
-
-  observeEvent(input$regressUI_subcheck, {
-    output$regressUI_subvar <- renderUI({
-      req(input$regressUI_subcheck == T)
-      factor_vars <- names(data)[data[, lapply(.SD, class) %in% c("factor", "character")]]
-      factor_subgroup <- setdiff(factor_vars, c(input$dep_vars, input$indep_vars))
-      if (!is.null(design.survey)){
-        factor_subgroup <- setdiff(factor_subgroup, c(names(design.survey$strata), names(design.survey$cluster)))
-      }
-      factor_subgroup_list <- mklist(data_varStruct, factor_subgroup)
-      validate(
-        need(length(factor_subgroup) > 0 , "No factor variable for sub-group analysis")
-      )
-
-      tagList(
-        selectInput(session$ns("subvar_logistic"), "Sub-group variable",
-                    choices = factor_subgroup_list, multiple = F,
-                    selected = factor_subgroup[1])
-      )
-    })
-
-    output$regressUI_subval <- renderUI({
-      req(input$regressUI_subcheck == T)
-      req(input$subvar_logistic)
-      selectInput(session$ns("subval_logistic"), "Sub-group value",
-                  choices = data_label[variable == input$subvar_logistic, level], multiple = F,
-                  selected = data_label[variable == input$subvar_logistic, level][1])
-    })
-  })
-
-
-
-
-
-  out <- reactive({
-    data.logistic <- data
-    if(input$regressUI_subcheck == T){
-      req(input$subvar_logistic)
-      data.logistic <- data.logistic[get(input$subvar_logistic) == input$subval_logistic, ]
-    }
-    y <- input$dep_vars
-    xs <- input$indep_vars
-    validate(
-      need(!is.null(input$indep_vars) , "Please select at least 1 variable")
-    )
-    form = as.formula(paste(y, "~", paste(xs, collapse = " + "), sep = " "))
-    mf <- model.frame(form, data.logistic)
-    validate(
-      need(nrow(mf) > 0, paste("No complete data due to missingness. Please remove some variables from independent variables"))
-    )
-    lgl.1level <- purrr::map_lgl(mf, ~length(unique(.x)) == 1)
-    validate(
-      need(sum(lgl.1level) == 0, paste(paste(names(lgl.1level)[lgl.1level], collapse =" ,"), "has(have) a unique value. Please remove that from independent variables"))
-    )
-
-    if (is.null(design.survey)){
-      res.logistic = glm(form, data = data.logistic, family = binomial)
-      tb.logistic = jstable::glmshow.display(res.logistic, decimal = input$decimal)
-      cap.logistic = paste("Logistic regression predicting ", data_label()[variable == y, var_label][1], sep="")
-      if(input$regressUI_subcheck == T){
-        cap.logistic <- paste(cap.logistic, " - ", data_label()[variable == input$subvar_logistic, var_label][1], ": ", data_label()[variable == input$subvar_logistic & level == input$subval_logistic, val_label], sep = "")
-      }
-      out.logistic = jstable::LabelepiDisplay(tb.logistic, label = T, ref = data_label())
-      #out.logistic = summary(res.logistic)$coefficients
-      #sig = ifelse(out.logistic[, 4] <= 0.05, "**", "NA")
-    } else{
-      #validate(
-      #  need(var.weights.survey %in% names(data.logistic) , "Weight variable isn't in data-, Please select appropriate weight variable.")
-      #)
-      #data.design <- survey::svydesign(ids = ~ 1, data = data.logistic, weights = ~ get(var.weights.survey))
-      data.design <- design.survey
-      if(input$regressUI_subcheck == T){
-        req(input$subvar_logistic)
-        data.design <- subset(data.design, get(input$subvar_logistic) == input$subval_logistic)
-      }
-
-      res.svyglm <- survey::svyglm(form, design = data.design, family = quasibinomial())
-      tb.svyglm <- jstable::svyregress.display(res.svyglm, decimal = input$decimal)
-      cap.logistic <- paste("Logistic regression predicting ", data_label()[variable == y, var_label][1], "- survey data", sep="")
-      if(input$regressUI_subcheck == T){
-        cap.logistic <- paste(cap.logistic, " - ", data_label()[variable == input$subvar_logistic, var_label][1], ": ", data_label()[variable == input$subvar_logistic & level == input$subval_regress, val_label], sep = "")
-      }
-      out.logistic <- jstable::LabelepiDisplay(tb.svyglm, label = T, ref = data_label())
-
-    }
-    return(list(table = out.logistic, caption = cap.logistic))
-  })
-
-  return(out)
-
-
-}
-
-
 
 
 
@@ -939,7 +525,7 @@ logisticModule <- function(input, output, session, data, data_label, data_varStr
 #' @import shiny
 #' @importFrom data.table data.table .SD :=
 #' @importFrom labelled var_label<-
-#' @importFrom stats glm as.formula binomial model.frame quasibinomial
+#' @importFrom stats glm as.formula binomial model.frame quasibinomial step
 #' @importFrom epiDisplay logistic.display
 #' @importFrom purrr map_lgl
 
@@ -947,7 +533,7 @@ logisticModule <- function(input, output, session, data, data_label, data_varStr
 logisticModule2 <- function(input, output, session, data, data_label, data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL, default.unires = T, limit.unires = 20) {
 
   ## To remove NOTE.
-  level <- val_label <- variable <- NULL
+  d2 <- level <- val_label <- variable <- NULL
 
   if (is.null(data_varStruct)){
     data_varStruct = reactive(list(variable = names(data())))
@@ -1093,6 +679,25 @@ logisticModule2 <- function(input, output, session, data, data_label, data_varSt
 
   })
 
+  observeEvent(input$step_check, {
+    output$step_direction <- renderUI({
+      req(input$step_check == T)
+      radioButtons(session$ns("step_direction"), "Step direction", choices = c("backward", "forward", "both"), selected = "backward", inline = T)
+    })
+
+    output$step_scope <- renderUI({
+      req(input$step_check == T)
+      req(input$indep_vars)
+      tagList(
+        fluidRow(
+          column(6, selectInput(session$ns("step_lower"), "Lower limit", choices = input$indep_vars, selected = NULL, multiple = T)),
+          column(6, selectInput(session$ns("step_upper"), "Upper limit", choices = input$indep_vars, selected = input$indep_vars, multiple = T))
+        )
+
+      )
+    })
+  })
+
 
 
 
@@ -1137,10 +742,22 @@ logisticModule2 <- function(input, output, session, data, data_label, data_varSt
 
 
     if (is.null(design.survey)){
-      res.logistic = glm(form, data = data.logistic, family = binomial)
-      tb.logistic = jstable::glmshow.display(res.logistic, decimal = input$decimal)
-      cap.logistic = paste("Logistic regression predicting ", label.regress[variable == y, var_label][1], sep="")
-      if(input$regressUI_subcheck == T){
+      res.logistic <- glm(form, data = data.logistic, family = binomial)
+      if (input$step_check == T){
+        validate(
+          need(!is.null(input$step_upper), "Upper limits can't be NULL, please select at least 1 variable."),
+          need(length((setdiff(input$step_lower, input$step_upper))) == 0, "Upper limits must include lower limits. Please add the variables to upper limits")
+        )
+        scope <- lapply(list(input$step_upper, input$step_lower), function(x){
+          as.formula(ifelse(is.null(x), "~1", paste0("~", paste(x, collapse = "+"))))
+        })
+
+        res.logistic <- stats::step(glm(form, data = data.logistic[complete.cases(data.logistic[, .SD, .SDcols = c(y, xs)])],  family = binomial), direction = input$step_direction, scope = list(upper = scope[[1]], lower = scope[[2]]))
+      }
+
+      tb.logistic <- jstable::glmshow.display(res.logistic, decimal = input$decimal)
+      cap.logistic <- paste("Logistic regression predicting ", label.regress[variable == y, var_label][1], sep="")
+      if (input$regressUI_subcheck == T){
         for (v in seq_along(input$subvar_logistic)){
           if (input$subvar_logistic[[v]] %in% vlist()$factor_vars){
             cap.logistic <- paste(cap.logistic, ", ", label.regress[variable == input$subvar_logistic[[v]], var_label][1], ": ", paste(label.regress[variable == input$subvar_logistic[[v]] & level %in% input[[paste0("subval_logistic", v)]], val_label], collapse = ", "), sep = "")
@@ -1149,7 +766,10 @@ logisticModule2 <- function(input, output, session, data, data_label, data_varSt
           }
         }
       }
-      out.logistic = jstable::LabelepiDisplay(tb.logistic, label = T, ref = label.regress)
+      if (input$step_check == T){
+        cap.logistic <- paste0(cap.logistic, "- stepwise selection")
+      }
+      out.logistic <- jstable::LabelepiDisplay(tb.logistic, label = T, ref = label.regress)
       #out.logistic = summary(res.logistic)$coefficients
       #sig = ifelse(out.logistic[, 4] <= 0.05, "**", "NA")
     } else{
@@ -1179,9 +799,24 @@ logisticModule2 <- function(input, output, session, data, data_label, data_varSt
       }
 
       res.svyglm <- survey::svyglm(form, design = data.design, family = quasibinomial())
+      if (input$step_check == T){
+        validate(
+          need(!is.null(input$step_upper), "Upper limits can't be NULL, please select at least 1 variable."),
+          need(length((setdiff(input$step_lower, input$step_upper))) == 0, "Upper limits must include lower limits. Please add the variables to upper limits"),
+          need(input$step_direction != "both", "Only forward/bacward directions are supported")
+        )
+        scope <- lapply(list(input$step_upper, input$step_lower), function(x){
+          as.formula(ifelse(is.null(x), "~1", paste0("~", paste(x, collapse = "+"))))
+        })
+
+        d2 <<- na.omit(data.design[, c(y, xs)])
+
+        res.svyglm <- stats::step(survey::svyglm(form, design = d2, family = quasibinomial()),
+                                  direction = input$step_direction, scope = list(upper = scope[[1]], lower = scope[[2]]))
+      }
       tb.svyglm <- jstable::svyregress.display(res.svyglm, decimal = input$decimal)
       cap.logistic <- paste("Logistic regression predicting ", label.regress[variable == y, var_label][1], "- survey data", sep="")
-      if(input$regressUI_subcheck == T){
+      if (input$regressUI_subcheck == T){
         for (v in seq_along(input$subvar_logistic)){
           if (input$subvar_logistic[[v]] %in% vlist()$factor_vars){
             cap.logistic <- paste(cap.logistic, ", ", label.regress[variable == input$subvar_logistic[[v]], var_label][1], ": ", paste(label.regress[variable == input$subvar_logistic[[v]] & level %in% input[[paste0("subval_logistic", v)]], val_label], collapse = ", "), sep = "")
@@ -1189,6 +824,9 @@ logisticModule2 <- function(input, output, session, data, data_label, data_varSt
             cap.logistic <- paste(cap.logistic, ", ", label.regress[variable == input$subvar_logistic[[v]], var_label][1], ": ", paste(input[[paste0("subval_logistic", v)]], collapse = "~"), sep = "")
           }
         }
+      }
+      if (input$step_check == T){
+        cap.logistic <- paste0(cap.logistic, "- stepwise selection")
       }
       out.logistic <- jstable::LabelepiDisplay(tb.svyglm, label = T, ref = label.regress)