+++ tests/testthat/test-module.R
@@ -23,11 +23,11 @@ test_that("Run Modules", {
 })
 
 
-context("basic")
+#context("basic")
 
-test_that("shinytest will be passed", {
-  skip_on_cran()
-  skip_on_travis()
-  shinytest::expect_pass(shinytest::testApp("shinytest/basic"))
-})
+#test_that("shinytest will be passed", {
+#  skip_on_cran()
+#  skip_on_travis()
+#  shinytest::expect_pass(shinytest::testApp("shinytest/basic"))
+#})